chore: create separate files for Desktop and Mobile view and bundle assets
