+++ erpnext/public/js/templates/node_card.html
