+++ erpnext/public/scss/hierarchy_chart.scss
